Move chat completions API endpoints to module
